requirements.txt
altair==6.0.0
antlr4-python3-runtime==4.9.3
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
decorator==5.2.1
demucs==4.0.1
dora_search==0.1.12
einops==0.8.1
filelock==3.20.1
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
julius==0.2.7
lameenc==1.8.1
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
moviepy==2.2.1
mpmath==1.3.0
narwhals==2.14.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
omegaconf==2.3.0
openai-whisper==20250625
openunmix==1.3.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
proglog==0.1.12
protobuf==6.33.2
pyarrow==22.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
retrying==1.4.2
rpds-py==0.30.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
streamlit==1.52.2
submitit==1.5.4
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
toml==0.10.2
torch==2.9.1
torchaudio==2.9.1
tornado==6.5.4
tqdm==4.67.1
treetable==0.2.6
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
watchdog==6.0.0
